+++ lib/cuprum/rails/records.rb
@@ -6,6 +6,7 @@ module Cuprum::Rails
   # The Records collection wraps ActiveRecord model persistence and querying.
   module Records
     autoload :Collection, 'cuprum/rails/records/collection'
+    autoload :Query,      'cuprum/rails/records/query'
     autoload :Repository, 'cuprum/rails/records/repository'
   end
 end